Make inlcude guard works as expected & Better Makefile!
