+++ lab3/bootloader/Makefile
@@ -10,7 +10,7 @@ LIB_SRC = \
 LIB_INCLUDES = \
 	-I $(LIB_SRC_DIR)/include \
 
-LIB_BUILD_DIR = $(BUILD_DIR)/lib
+LIB_BUILD_DIR = $(LIB_SRC_DIR)/build/lib
 LIB_OBJS := $(patsubst %.c,$(LIB_BUILD_DIR)/%.o,$(notdir $(LIB_SRC)))
 
 # Application
@@ -60,7 +60,7 @@ LINKER_FILE = linker.ld
 CCFLAGS = -Wall -O0 -fno-stack-protector
 LDFLAGS = -T$(LINKER_FILE) -nostdlib
 
-.PHONY: clean directories out_elf run docker_cp
+.PHONY: clean directories out_elf run
 all: directories $(OUT_OBJS) $(OUT_ELF) $(OUT_IMG)
 directories: $(BUILD_DIRECTORIES)
 out_elf: directories $(OUT_ELF)
@@ -73,7 +73,7 @@ $(APP_BUILD_DIR)/%.o : %.c
 	@echo '============================================================'
 	@echo "[+] Building: $<"
 	$(CC) $(CCFLAGS) $(APP_INCLUDES) -g -o $@ -c $<
-	@echo "[+] Finished Building: $<"
+	@echo "[+] Finish Building: $<"
 	@echo ''
 
 $(APP_BUILD_DIR)/%.o : %.S
@@ -82,16 +82,7 @@ $(APP_BUILD_DIR)/%.o : %.S
 	@echo '============================================================'
 	@echo "[+] Building: $<"
 	$(CC) $(ASMFLAGS) $(APP_INCLUDES) -g -o $@ -c $<
-	@echo "[+] Finished Building: $<"
-	@echo ''
-
-$(LIB_BUILD_DIR)/%.o : %.c
-	@echo '============================================================'
-	@echo ' Building target: $@'
-	@echo '============================================================'
-	@echo "[+] Building: $<"
-	$(CC) $(CCFLAGS) $(LIB_INCLUDES) -g -o $@ -c $<
-	@echo "[+] Finished Building: $<"
+	@echo "[+] Finish Building: $<"
 	@echo ''
 
 # Generate ELF -----------------------------------------------------------------
@@ -102,7 +93,7 @@ $(OUT_ELF): $(OUT_OBJS)
 	$(LD) $(LDFLAGS) -o $@ $(OUT_OBJS)
 	$(OD) -d $@ > $(BUILD_DIR)/$(PROJ).objdump
 	# $(SP) $@
-	@echo '[+] Finished building target: $@'
+	@echo '[+] Finish Building target: $@'
 	@echo ' '
 
 # Generate IMG -----------------------------------------------------------------
@@ -111,7 +102,7 @@ $(OUT_IMG): $(OUT_ELF)
 	@echo ' Building target: $@'
 	@echo '============================================================'
 	$(OC) -O binary $(OUT_ELF) $@
-	@echo '[+] Finished building target: $@'
+	@echo '[+] Finish Building target: $@'
 	@echo ' '
 
 $(BUILD_DIRECTORIES):
