+++ lab3/kernel/Makefile
@@ -7,6 +7,8 @@ LIB_SRC = \
 	$(LIB_SRC_DIR)/malloc.c \
 	$(LIB_SRC_DIR)/mmio.c \
 	$(LIB_SRC_DIR)/mbox.c \
+	$(LIB_SRC_DIR)/reset.c \
+	$(LIB_SRC_DIR)/string.c \
 	$(LIB_SRC_DIR)/shell.c \
 	$(LIB_SRC_DIR)/uart.c \
 	$(LIB_SRC_DIR)/dtb.c \
@@ -14,7 +16,7 @@ LIB_SRC = \
 LIB_INCLUDES = \
 	-I $(LIB_SRC_DIR)/include \
 
-LIB_BUILD_DIR = $(BUILD_DIR)/lib
+LIB_BUILD_DIR = $(LIB_SRC_DIR)/build/lib
 LIB_OBJS := $(patsubst %.c,$(LIB_BUILD_DIR)/%.o,$(notdir $(LIB_SRC)))
 
 # Application
@@ -34,9 +36,6 @@ APP_BUILD_DIR = $(BUILD_DIR)/app
 APP_OBJS := $(patsubst %.c,$(APP_BUILD_DIR)/%.o,$(notdir $(APP_SRC)))
 APP_OBJS += $(patsubst %.S,$(APP_BUILD_DIR)/%.o,$(notdir $(APP_ASM)))
 
-VPATH += \
-	$(LIB_SRC_DIR) \
-	$(APP_SRC_DIR) \
 
 # ARM toolchain
 CROSS = aarch64-linux-gnu
@@ -64,8 +63,8 @@ LINKER_FILE = linker.ld
 CCFLAGS = -Wall -O0 -fno-stack-protector
 LDFLAGS = -T$(LINKER_FILE) -nostdlib -nostartfiles
 
-.PHONY: clean directories out_elf run docker_cp
-all: directories $(OUT_OBJS) $(OUT_ELF) $(OUT_IMG)
+.PHONY: clean directories out_elf run
+all: directories $(OUT_ELF) $(OUT_IMG)
 directories: $(BUILD_DIRECTORIES)
 out_elf: directories $(OUT_ELF)
 out_img: directories $(OUT_IMG)
@@ -77,7 +76,7 @@ $(APP_BUILD_DIR)/%.o : %.c
 	@echo '============================================================'
 	@echo "[+] Building: $<"
 	$(CC) $(CCFLAGS) $(APP_INCLUDES) -g -o $@ -c $<
-	@echo "[+] Finished Building: $<"
+	@echo "[+] Finish Building: $<"
 	@echo ''
 
 $(APP_BUILD_DIR)/%.o : %.S
@@ -86,16 +85,7 @@ $(APP_BUILD_DIR)/%.o : %.S
 	@echo '============================================================'
 	@echo "[+] Building: $<"
 	$(CC) $(ASMFLAGS) $(APP_INCLUDES) -g -o $@ -c $<
-	@echo "[+] Finished Building: $<"
-	@echo ''
-
-$(LIB_BUILD_DIR)/%.o : %.c
-	@echo '============================================================'
-	@echo ' Building target: $@'
-	@echo '============================================================'
-	@echo "[+] Building: $<"
-	$(CC) $(CCFLAGS) $(LIB_INCLUDES) -g -o $@ -c $<
-	@echo "[+] Finished Building: $<"
+	@echo "[+] Finish Building: $<"
 	@echo ''
 
 # Generate ELF -----------------------------------------------------------------
@@ -106,7 +96,7 @@ $(OUT_ELF): $(OUT_OBJS)
 	$(LD) $(LDFLAGS) -o $@ $(OUT_OBJS)
 	$(OD) -d $@ > $(BUILD_DIR)/$(PROJ).objdump
 	# $(SP) $@
-	@echo '[+] Finished building target: $@'
+	@echo '[+] Finish Building target: $@'
 	@echo ' '
 
 # Generate IMG -----------------------------------------------------------------
@@ -115,30 +105,12 @@ $(OUT_IMG): $(OUT_ELF)
 	@echo ' Building target: $@'
 	@echo '============================================================'
 	$(OC) -O binary $(OUT_ELF) $@
-	@echo '[+] Finished building target: $@'
+	@echo '[+] Finish Building target: $@'
 	@echo ' '
 
 $(BUILD_DIRECTORIES):
 	mkdir -p $@
 
-run:
-	qemu-system-aarch64 -M raspi3 -kernel bootloader.img -display none -serial null -serial pty 
-
-run_stdio:
-	qemu-system-aarch64 -M raspi3 -kernel bootloader.img -display none -serial null -serial stdio
-
-run_cpio:
-	qemu-system-aarch64 -M raspi3 -kernel kernel.img -display none -serial null -serial stdio -initrd /root/osc2024/lab2/initramfs.cpio -dtb /root/osc2024/lab2/bcm2710-rpi-3-b-plus.dtb
-
-debug:
-	qemu-system-aarch64 -M raspi3 -kernel bootloader.img -display none -serial null -serial pty -s -S
-
-docker_cp:
-	docker cp kernel.elf my_kali:/
-	docker cp bootloader/bootloader.elf my_kali:/
-
-cpio:
-	cd /root/osc2024/lab2/rootfs && find . | cpio -o -H newc > ../initramfs.cpio
 
 clean:
 	rm -rf *.elf
