+++ lab3/lib/include/malloc.h
@@ -1,4 +1,4 @@
-#ifndef __MEM__H__
+#ifndef __MEM_H__
 #define __MEM_H__
 
 #include "stdint.h"
